requirements.txt
streamlit==1.32.0
sqlalchemy==2.0.27
psycopg[binary]==3.1.18
python-dotenv==1.0.1
alembic==1.13.1
pandas==2.2.0
plotly==5.18.0
bcrypt==4.1.2
python-jose==3.3.0
pytest==8.0.0
python-multipart==0.0.9 